# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conversational AI Dashboard Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardQueryRequest {
  query: string;
  /** @nullable */
  tableContext?: string | null;
}

export type DashboardQueryResponseDataItem = { [key: string]: unknown };

export type ChartConfigType =
  (typeof ChartConfigType)[keyof typeof ChartConfigType];

export const ChartConfigType = {
  line: "line",
  bar: "bar",
  pie: "pie",
  horizontalBar: "horizontalBar",
  table: "table",
} as const;

export type ChartConfigDataItem = { [key: string]: unknown };

export interface ChartConfig {
  type: ChartConfigType;
  title: string;
  /** @nullable */
  xKey?: string | null;
  /** @nullable */
  yKey?: string | null;
  dataKeys: string[];
  data: ChartConfigDataItem[];
  columns?: string[];
}

export interface DashboardQueryResponse {
  sql: string;
  data: DashboardQueryResponseDataItem[];
  chartConfigs: ChartConfig[];
  insights: string[];
  cannotAnswer: boolean;
  rowCount: number;
}

export interface UploadCsvRequest {
  tableName: string;
  csvContent: string;
}

export interface UploadCsvResponse {
  success: boolean;
  tableName: string;
  rowCount: number;
  columns: string[];
}

export interface TableInfo {
  name: string;
  rowCount: number;
  columns: string[];
}

export interface SchemaInfo {
  tables: TableInfo[];
}

export interface QueryHistoryItem {
  id: number;
  query: string;
  sql: string;
  createdAt: string;
}

export interface ApiError {
  error: string;
}