# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conversational AI Dashboard Generator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Query the dashboard with natural language
 */
export const QueryDashboardBody = zod.object({
  query: zod.string(),
  tableContext: zod.string().nullish(),
});

export const QueryDashboardResponse = zod.object({
  sql: zod.string(),
  data: zod.array(zod.record(zod.string(), zod.unknown())),
  chartConfigs: zod.array(
    zod.object({
      type: zod.enum(["line", "bar", "pie", "horizontalBar", "table"]),
      title: zod.string(),
      xKey: zod.string().nullish(),
      yKey: zod.string().nullish(),
      dataKeys: zod.array(zod.string()),
      data: zod.array(zod.record(zod.string(), zod.unknown())),
      columns: zod.array(zod.string()).optional(),
    }),
  ),
  insights: zod.array(zod.string()),
  cannotAnswer: zod.boolean(),
  rowCount: zod.number(),
});

/**
 * @summary Upload a CSV file and store in database
 */
export const UploadCsvBody = zod.object({
  tableName: zod.string(),
  csvContent: zod.string(),
});

export const UploadCsvResponse = zod.object({
  success: zod.boolean(),
  tableName: zod.string(),
  rowCount: zod.number(),
  columns: zod.array(zod.string()),
});

/**
 * @summary List available data tables
 */
export const ListTablesResponseItem = zod.object({
  name: zod.string(),
  rowCount: zod.number(),
  columns: zod.array(zod.string()),
});
export const ListTablesResponse = zod.array(ListTablesResponseItem);

/**
 * @summary Get schema information for available tables
 */
export const GetSchemaInfoResponse = zod.object({
  tables: zod.array(
    zod.object({
      name: zod.string(),
      rowCount: zod.number(),
      columns: zod.array(zod.string()),
    }),
  ),
});

/**
 * @summary Get recent query history
 */
export const GetQueryHistoryResponseItem = zod.object({
  id: zod.number(),
  query: zod.string(),
  sql: zod.string(),
  createdAt: zod.date(),
});
export const GetQueryHistoryResponse = zod.array(GetQueryHistoryResponseItem);