# lib/api-zod/src/generated/types/chartConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conversational AI Dashboard Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartConfigDataItem } from "./chartConfigDataItem";
import type { ChartConfigType } from "./chartConfigType";

export interface ChartConfig {
  type: ChartConfigType;
  title: string;
  /** @nullable */
  xKey?: string | null;
  /** @nullable */
  yKey?: string | null;
  dataKeys: string[];
  data: ChartConfigDataItem[];
  columns?: string[];
}