# lib/api-zod/src/generated/types/dashboardQueryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conversational AI Dashboard Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardQueryRequest {
  query: string;
  /** @nullable */
  tableContext?: string | null;
}