# lib/api-zod/src/generated/types/dashboardQueryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conversational AI Dashboard Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartConfig } from "./chartConfig";
import type { DashboardQueryResponseDataItem } from "./dashboardQueryResponseDataItem";

export interface DashboardQueryResponse {
  sql: string;
  data: DashboardQueryResponseDataItem[];
  chartConfigs: ChartConfig[];
  insights: string[];
  cannotAnswer: boolean;
  rowCount: number;
}