# lib/api-zod/src/generated/types/schemaInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conversational AI Dashboard Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { TableInfo } from "./tableInfo";

export interface SchemaInfo {
  tables: TableInfo[];
}